requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-mail==0.10.0
flask-wtf==1.2.1
python-dotenv==1.0.0
pillow==10.2.0
razorpay==1.4.1
sqlalchemy==2.0.25
werkzeug==3.0.1
gunicorn==21.2.0
requests==2.31.0
emoji==2.10.1
